pubsub_to_sqs/requirements.txt
google-cloud-pubsub
functions-framework
boto3
google-cloud-logging